backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pillow>=10.3
ultralytics==8.1.0
numpy>=1.26
gradio>=4.27
requests>=2.31
boto3>=1.34
